Airborne Threats Monitoring and Oxygen Filtration System
Remake of AIRDUINO (2020). It has the same features with upgraded designs. It can monitor at the same time filter harmful particulates presents in the air with the help of HEPA Filter. There is also an OLED screen that displays the status of air around the device. Also, there's a light notification and buzz alert. This device also has its own SD Card that can store data at the same time it can be connected to the internet since it is equipped with Wi-Fi Module.